cbify refactor
