+++ vowpalwabbit/cb_algs.cc
@@ -9,27 +9,16 @@ license as described in the file LICENSE.
 #include "reductions.h"
 #include "cb_algs.h"
 #include "vw_exception.h"
+#include "gen_cs_example.h"
 
 using namespace LEARNER;
 
-#define CB_TYPE_DR 0
-#define CB_TYPE_DM 1
-#define CB_TYPE_IPS 2
-
 using namespace CB;
 
-struct cb
-{ size_t cb_type;
-  uint32_t num_actions;
+struct cb 
+{ 
+  cb_to_cs& cbcs;
   COST_SENSITIVE::label cb_cs_ld;
-  COST_SENSITIVE::label pred_scores;
-  LEARNER::base_learner* scorer;
-  float avg_loss_regressors;
-  size_t nb_ex_regressors;
-  float last_pred_reg;
-  float last_correct_cost;
-
-  cb_class* known_cost;
 };
 
 bool know_all_cost_example(CB::label& ld)
@@ -54,215 +43,20 @@ bool is_test_label(CB::label& ld)
   return true;
 }
 
-inline bool observed_cost(cb_class* cl)
-{ //cost observed for this action if it has non zero probability and cost != FLT_MAX
-  return (cl != nullptr && cl->cost != FLT_MAX && cl->probability > .0);
-}
-
-cb_class* get_observed_cost(CB::label& ld)
-{ for (auto& cl : ld.costs)
-    if( observed_cost(&cl) )
-      return &cl;
-  return nullptr;
-}
-
-void gen_cs_example_ips(cb& c, CB::label& ld, COST_SENSITIVE::label& cs_ld)
-{ //this implements the inverse propensity score method, where cost are importance weighted by the probability of the chosen action
-  //generate cost-sensitive example
-  cs_ld.costs.erase();
-  if( ld.costs.size() == 1)   //this is a typical example where we can perform all actions
-  { //in this case generate cost-sensitive example with all actions
-    for(uint32_t i = 1; i <= c.num_actions; i++)
-    { COST_SENSITIVE::wclass wc;
-      wc.wap_value = 0.;
-      wc.x = 0.;
-      wc.class_index = i;
-      wc.partial_prediction = 0.;
-      wc.wap_value = 0.;
-      if( c.known_cost != nullptr && i == c.known_cost->action )
-      { wc.x = c.known_cost->cost / c.known_cost->probability; //use importance weighted cost for observed action, 0 otherwise
-        //ips can be thought as the doubly robust method with a fixed regressor that predicts 0 costs for everything
-        //update the loss of this regressor
-        c.nb_ex_regressors++;
-        c.avg_loss_regressors += (1.0f/c.nb_ex_regressors)*( (c.known_cost->cost)*(c.known_cost->cost) - c.avg_loss_regressors );
-        c.last_pred_reg = 0;
-        c.last_correct_cost = c.known_cost->cost;
-      }
-
-      cs_ld.costs.push_back(wc );
-    }
-  }
-  else   //this is an example where we can only perform a subset of the actions
-  { //in this case generate cost-sensitive example with only allowed actions
-    for (auto& cl : ld.costs)
-    { COST_SENSITIVE::wclass wc;
-      wc.wap_value = 0.;
-      wc.x = 0.;
-      wc.class_index = cl.action;
-      wc.partial_prediction = 0.;
-      wc.wap_value = 0.;
-      if( c.known_cost != nullptr && cl.action == c.known_cost->action )
-      { wc.x = c.known_cost->cost / c.known_cost->probability; //use importance weighted cost for observed action, 0 otherwise
-
-        //ips can be thought as the doubly robust method with a fixed regressor that predicts 0 costs for everything
-        //update the loss of this regressor
-        c.nb_ex_regressors++;
-        c.avg_loss_regressors += (1.0f/c.nb_ex_regressors)*( (c.known_cost->cost)*(c.known_cost->cost) - c.avg_loss_regressors );
-        c.last_pred_reg = 0;
-        c.last_correct_cost = c.known_cost->cost;
-      }
-
-      cs_ld.costs.push_back( wc );
-    }
-  }
-
-}
 
 template <bool is_learn>
-void gen_cs_example_dm(cb& c, example& ec, COST_SENSITIVE::label& cs_ld)
-{ //this implements the direct estimation method, where costs are directly specified by the learned regressor.
-  CB::label ld = ec.l.cb;
-
-  float min = FLT_MAX;
-  uint32_t argmin = 1;
-  //generate cost sensitive example
-  cs_ld.costs.erase();
-  c.pred_scores.costs.erase();
-
-  if( ld.costs.size() == 1)   //this is a typical example where we can perform all actions
-  { //in this case generate cost-sensitive example with all actions
-    for(uint32_t i = 1; i <= c.num_actions; i++)
-    { COST_SENSITIVE::wclass wc;
-      wc.wap_value = 0.;
-
-      //get cost prediction for this action
-      wc.x = CB_ALGS::get_cost_pred<is_learn>(c.scorer, c.known_cost, ec, i, 0);
-      if (wc.x < min)
-      { min = wc.x;
-        argmin = i;
-      }
-
-      wc.class_index = i;
-      wc.partial_prediction = 0.;
-      wc.wap_value = 0.;
-
-      c.pred_scores.costs.push_back(wc);
-
-      if( c.known_cost != nullptr && c.known_cost->action == i )
-      { c.nb_ex_regressors++;
-        c.avg_loss_regressors += (1.0f/c.nb_ex_regressors)*( (c.known_cost->cost - wc.x)*(c.known_cost->cost - wc.x) - c.avg_loss_regressors );
-        c.last_pred_reg = wc.x;
-        c.last_correct_cost = c.known_cost->cost;
-      }
-
-      cs_ld.costs.push_back( wc );
-    }
-  }
-  else   //this is an example where we can only perform a subset of the actions
-  { //in this case generate cost-sensitive example with only allowed actions
-    for (auto& cl : ld.costs)
-    { COST_SENSITIVE::wclass wc;
-      wc.wap_value = 0.;
-
-      //get cost prediction for this action
-      wc.x = CB_ALGS::get_cost_pred<is_learn>(c.scorer, c.known_cost, ec, cl.action, 0);
-      if (wc.x < min || (wc.x == min && cl.action < argmin))
-      { min = wc.x;
-        argmin = cl.action;
-      }
-
-      wc.class_index = cl.action;
-      wc.partial_prediction = 0.;
-      wc.wap_value = 0.;
-
-      c.pred_scores.costs.push_back(wc);
-
-      if( c.known_cost != nullptr && c.known_cost->action == cl.action )
-      { c.nb_ex_regressors++;
-        c.avg_loss_regressors += (1.0f/c.nb_ex_regressors)*( (c.known_cost->cost - wc.x)*(c.known_cost->cost - wc.x) - c.avg_loss_regressors );
-        c.last_pred_reg = wc.x;
-        c.last_correct_cost = c.known_cost->cost;
-      }
-
-      cs_ld.costs.push_back( wc );
-    }
-  }
-
-  ec.pred.multiclass = argmin;
-}
-
-template <bool is_learn>
-void gen_cs_label(cb& c, example& ec, COST_SENSITIVE::label& cs_ld, uint32_t label)
-{ COST_SENSITIVE::wclass wc;
-  wc.wap_value = 0.;
-
-  //get cost prediction for this label
-  wc.x = CB_ALGS::get_cost_pred<is_learn>(c.scorer, c.known_cost, ec, label, c.num_actions);
-  wc.class_index = label;
-  wc.partial_prediction = 0.;
-  wc.wap_value = 0.;
-
-  c.pred_scores.costs.push_back(wc);
-
-  //add correction if we observed cost for this action and regressor is wrong
-  if( c.known_cost != nullptr && c.known_cost->action == label )
-  { c.nb_ex_regressors++;
-    c.avg_loss_regressors += (1.0f/c.nb_ex_regressors)*( (c.known_cost->cost - wc.x)*(c.known_cost->cost - wc.x) - c.avg_loss_regressors );
-    c.last_pred_reg = wc.x;
-    c.last_correct_cost = c.known_cost->cost;
-    wc.x += (c.known_cost->cost - wc.x) / c.known_cost->probability;
-  }
-  //cout<<"Prediction = "<<wc.x<<" ";
-  cs_ld.costs.push_back( wc );
-
-}
-
-template <bool is_learn>
-void gen_cs_example_dr(cb& c, example& ec, CB::label& ld, COST_SENSITIVE::label& cs_ld)
-{ //this implements the doubly robust method
-  cs_ld.costs.erase();
-  c.pred_scores.costs.erase();
-  if(ld.costs.size() == 0)//a test example
-    for(uint32_t i = 1; i <= c.num_actions; i++)
-    { //Explicit declaration for a weak compiler.
-      COST_SENSITIVE::wclass c = {FLT_MAX, i, 0., 0.};
-      cs_ld.costs.push_back(c);
-    }
-  else if( ld.costs.size() == 1) //this is a typical example where we can perform all actions
-    //in this case generate cost-sensitive example with all actions
-    for(uint32_t i = 1; i <= c.num_actions; i++)
-      gen_cs_label<is_learn>(c, ec, cs_ld, i);
-  else  //this is an example where we can only perform a subset of the actions
-    //in this case generate cost-sensitive example with only allowed actions
-    for (auto& cl : ld.costs)
-      gen_cs_label<is_learn>(c, ec, cs_ld, cl.action);
-  //cout<<endl;
-}
-
-template <bool is_learn>
-void predict_or_learn(cb& c, base_learner& base, example& ec)
+void predict_or_learn(cb& data, base_learner& base, example& ec)
 { CB::label ld = ec.l.cb;
 
+  cb_to_cs& c = data.cbcs;
   c.known_cost = get_observed_cost(ld);
   if (c.known_cost != nullptr && (c.known_cost->action < 1 || c.known_cost->action > c.num_actions))
     cerr << "invalid action: " << c.known_cost->action << endl;
   //generate a cost-sensitive example to update classifiers
-  switch(c.cb_type)
-  { case CB_TYPE_IPS:
-      gen_cs_example_ips(c,ld,c.cb_cs_ld);
-      break;
-    case CB_TYPE_DM:
-      gen_cs_example_dm<is_learn>(c,ec,c.cb_cs_ld);
-      break;
-    case CB_TYPE_DR:
-      gen_cs_example_dr<is_learn>(c,ec,ld,c.cb_cs_ld);
-      break;
-    default:
-      THROW("Unknown cb_type specified for contextual bandit learning: " << c.cb_type);
-  }
+  gen_cs_example<is_learn>(c, ec, ld, data.cb_cs_ld);
 
   if (c.cb_type != CB_TYPE_DM)
-  { ec.l.cs = c.cb_cs_ld;
+  { ec.l.cs = data.cb_cs_ld;
 
     if (is_learn)
       base.learn(ec);
@@ -270,7 +64,7 @@ void predict_or_learn(cb& c, base_learner& base, example& ec)
       base.predict(ec);
 
     for (size_t i=0; i<ld.costs.size(); i++)
-      ld.costs[i].partial_prediction = c.cb_cs_ld.costs[i].partial_prediction;
+      ld.costs[i].partial_prediction = data.cb_cs_ld.costs[i].partial_prediction;
     ec.l.cb = ld;
   }
 }
@@ -279,18 +73,15 @@ void predict_eval(cb&, base_learner&, example&)
 { THROW("can not use a test label for evaluation");
 }
 
-void learn_eval(cb& c, base_learner&, example& ec)
+void learn_eval(cb& data, base_learner&, example& ec)
 { CB_EVAL::label ld = ec.l.cb_eval;
 
+  cb_to_cs& c = data.cbcs;
   c.known_cost = get_observed_cost(ld.event);
-
-  if (c.cb_type == CB_TYPE_DR)
-    gen_cs_example_dr<true>(c, ec, ld.event, c.cb_cs_ld);
-  else //c.cb_type == CB_TYPE_IPS
-    gen_cs_example_ips(c, ld.event, c.cb_cs_ld);
+  gen_cs_example<true>(c, ec, ld.event, data.cb_cs_ld);
 
   for (size_t i=0; i<ld.event.costs.size(); i++)
-    ld.event.costs[i].partial_prediction = c.cb_cs_ld.costs[i].partial_prediction;
+    ld.event.costs[i].partial_prediction = data.cb_cs_ld.costs[i].partial_prediction;
 
   ec.pred.multiclass = ec.l.cb_eval.action;
 }
@@ -303,8 +94,10 @@ float get_unbiased_cost(CB::cb_class* observation, COST_SENSITIVE::label& scores
   return get_unbiased_cost(observation, action);
 }
 
-void output_example(vw& all, cb& c, example& ec, CB::label& ld)
+void output_example(vw& all, cb& data, example& ec, CB::label& ld)
 { float loss = 0.;
+
+  cb_to_cs& c = data.cbcs;
   if(!is_test_label(ld))
     loss = get_unbiased_cost(c.known_cost, c.pred_scores, ec.pred.multiclass);
 
@@ -326,8 +119,13 @@ void output_example(vw& all, cb& c, example& ec, CB::label& ld)
   print_update(all, is_test_label(ld), ec, nullptr, false);
 }
 
-void finish(cb& c)
-{ c.cb_cs_ld.costs.delete_v(); c.pred_scores.costs.delete_v();}
+void finish(cb& data)
+{ 
+  cb_to_cs& c = data.cbcs;
+  data.cb_cs_ld.costs.delete_v(); 
+  COST_SENSITIVE::cs_label.delete_label(&c.pred_scores); 
+  free(&c);
+}
 
 void finish_example(vw& all, cb& c, example& ec)
 { output_example(all, c, ec, ec.l.cb);
@@ -347,7 +145,9 @@ base_learner* cb_algs_setup(vw& all)
   ("eval", "Evaluate a policy rather than optimizing.");
   add_options(all);
 
-  cb& c = calloc_or_throw<cb>();
+  cb& data = calloc_or_throw<cb>();
+  data.cbcs = calloc_or_throw<cb_to_cs>();
+  cb_to_cs& c = data.cbcs;
   c.num_actions = (uint32_t)all.vm["cb"].as<size_t>();
 
   bool eval = false;
@@ -402,11 +202,11 @@ base_learner* cb_algs_setup(vw& all)
 
   learner<cb>* l;
   if (eval)
-  { l = &init_learner(&c, base, learn_eval, predict_eval, problem_multiplier);
+  { l = &init_learner(&data, base, learn_eval, predict_eval, problem_multiplier);
     l->set_finish_example(eval_finish_example);
   }
   else
-  { l = &init_learner(&c, base, predict_or_learn<true>, predict_or_learn<false>,
+  { l = &init_learner(&data, base, predict_or_learn<true>, predict_or_learn<false>,
                       problem_multiplier);
     l->set_finish_example(finish_example);
   }
