+++ vowpalwabbit/cb_algs.h
@@ -7,6 +7,10 @@ license as described in the file LICENSE.
 //TODO: extend to handle CSOAA_LDF and WAP_LDF
 LEARNER::base_learner* cb_algs_setup(vw& all);
 
+#define CB_TYPE_DR 0
+#define CB_TYPE_DM 1
+#define CB_TYPE_IPS 2
+
 namespace CB_ALGS
 {
 template <bool is_learn>
