+++ vowpalwabbit/cb_explore.h
@@ -2,6 +2,8 @@
 Copyright (c) by respective owners including Yahoo!, Microsoft, and
 individual contributors. All rights reserved.  Released under a BSD
 license as described in the file LICENSE.
- */
-LEARNER::base_learner* cbify_setup(vw& all);
+*/
+#pragma once
+
+LEARNER::base_learner* cb_explore_setup(vw& all);
