+++ vowpalwabbit/vw_dynamic.vcxproj
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project DefaultTargets="Build" ToolsVersion="12.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup Label="ProjectConfigurations">
     <ProjectConfiguration Include="DebugLeakCheck|Win32">
@@ -345,7 +345,7 @@
       <FileType>CppHeader</FileType>
     </ClCompile>
     <ClInclude Include="active_cover.h" />
-    <ClCompile Include="mwt.cc" />
+    <!-- <ClCompile Include="mwt.cc" /> -->
     <ClCompile Include="primitives.cc" />
     <ClCompile Include="vw_validate.cc" />
     <ClInclude Include="autolink.h" />
@@ -355,7 +355,10 @@
     <ClInclude Include="bfgs.h" />
     <ClInclude Include="binary.h" />
     <ClInclude Include="cache.h" />
+    <ClInclude Include="cb.h" />
+    <ClInclude Include="cbify.h" />
     <ClInclude Include="cb_adf.h" />
+    <ClInclude Include="cb_explore.h" />
     <ClInclude Include="comp_io.h" />
     <ClInclude Include="confidence.h" />
     <ClInclude Include="constant.h" />
@@ -365,6 +368,7 @@
     <ClInclude Include="example.h" />
     <ClInclude Include="feature_group.h" />
     <ClInclude Include="gd.h" />
+    <ClInclude Include="gen_cs_example.h" />
     <ClInclude Include="interactions.h" />
     <ClInclude Include="audit_regressor.h" />
     <ClInclude Include="ftrl.h" />
@@ -388,7 +392,7 @@
     <ClInclude Include="loss_functions.h" />
     <ClInclude Include="multilabel.h" />
     <ClInclude Include="multilabel_oaa.h" />
-    <ClInclude Include="mwt.h" />
+    <!-- <ClInclude Include="mwt.h" /> -->
     <ClInclude Include="network.h" />
     <ClInclude Include="nn.h" />
     <ClInclude Include="noop.h" />
@@ -440,6 +444,7 @@
     <ClCompile Include="cb.cc" />
     <ClCompile Include="cbify.cc" />
     <ClCompile Include="cb_adf.cc" />
+    <ClCompile Include="cb_explore.cc" />
     <ClCompile Include="comp_io.cc" />
     <ClCompile Include="confidence.cc" />
     <ClCompile Include="csoaa.cc" />
@@ -457,6 +462,7 @@
     <ClCompile Include="cb_algs.cc" />
     <ClCompile Include="mwt.cc" />
     <ClCompile Include="mf.cc" />
+    <ClCompile Include="gen_cs_example.cc" />
     <ClCompile Include="gd_mf.cc" />
     <ClCompile Include="lrq.cc" />
     <ClCompile Include="lrqfa.cc" />
